tests/wpt/tests/websockets/opening-handshake/001.html

<!doctype html>
<title>WebSockets: invalid handshake</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.sub.js></script>
<meta name="variant" content="?default">
<meta name="variant" content="?wss">
<meta name="variant" content="?wpt_flags=h2">
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/invalid');
  ws.onclose = t.step_func(function(e) {
    assert_false(e.wasClean);
    ws.onclose = t.unreached_func();
    t.step_timeout(() => t.done(), 50);
  });
   ws.onmessage = ws.onopen = t.unreached_func();
});
</script>
